Freq in MHz

+++ rf_mbus.hpp
@@ -82,7 +82,7 @@ const std::string mode_to_string(WmBusFrameMode mode);
 class rf_mbus {
   public:
     bool init(uint8_t mosi, uint8_t miso, uint8_t clk, uint8_t cs,
-              uint8_t gdo0, uint8_t gdo2, uint32_t freq);
+              uint8_t gdo0, uint8_t gdo2, float freq);
     bool task();
     WMbusFrame get_frame();
